+++ tasmoadmin/composer.json
@@ -8,10 +8,9 @@
         "philipp15b/php-i18n": "^4.0"
     },
     "autoload": {
-        "files": [
-            "helpers/LoginHelper.php",
-            "includes/Config.php"
-        ]
+        "psr-4": {
+            "TasmoAdmin\\": "src/"
+        }
     },
     "require-dev": {
         "phpunit/phpunit": "^9.3.8"
