+++ tasmoadmin/pages/device_update.php
@@ -1,5 +1,7 @@
 <?php
 //	var_dump( $_REQUEST );
+use TasmoAdmin\Helper\UrlHelper;
+
 $useSSL    = $Config->read("ota_server_ssl");
 $localIP   = $Config->read("ota_server_ip");
 $localPort = $Config->read("ota_server_port");
