+++ tasmoadmin/pages/devices.php
@@ -1,5 +1,7 @@
 <?php
 
+use TasmoAdmin\Helper\UrlHelper;
+
 $devices = $Sonoff->getDevices();
