+++ tasmoadmin/pages/selfupdate.php
@@ -1,6 +1,8 @@
 <?php
 
-require_once _APPROOT_ . "vendor/autoload.php"; 
+use TasmoAdmin\Selfupdate;
+
+require_once _APPROOT_ . "vendor/autoload.php";
 
 require_once _INCLUDESDIR_ . "Selfupdate.php";
