+++ tasmoadmin/src/Helper/LoginHelper.php
@@ -1,17 +1,20 @@
 <?php
 
+namespace TasmoAdmin\Helper;
+
+use TasmoAdmin\Config;
+
 class LoginHelper
 {
-    
+    private Config $config;
+
     public function __construct(Config $config)
     {
         $this->config = $config;
     }
 
-    public function login($password, $storedPassword)
+    public function login($password, $storedPassword): bool
     {
-        $passwordHashed = password_hash($password, PASSWORD_DEFAULT);
-
         //update hashing
         if (strpos($storedPassword, '$2y$') !== 0) {
             if ($storedPassword !== md5($password)) {
@@ -32,13 +35,13 @@ class LoginHelper
         return true;
     }
 
-    public function register($username, $password)
+    public function register($username, $password): void
     {
         $this->config->write("username", $username);
         $this->config->write("password", self::hashPassword($password));
     }
 
-    private static function hashPassword($password)
+    private static function hashPassword($password): string
     {
         return password_hash($password, PASSWORD_DEFAULT);
     }
